Add CORS support
Solves the CORS issue for the AWS CDK API Gateway and Lambda Backend.
A simple website with an HTML and JS file was created to test this.
JS file includes a fetch function that makes a GET request to the specified URL. Fetched data is parsed as JSON and logged to the console.